[db][cmd] Move commands into seperate files
- previous it's just `cmd.go`, now it's a `cmd` package
