+++ ice/db/cmd/config.go
@@ -1,0 +1,15 @@
+package cmd
+
+import "github.com/spf13/cobra"
+
+func makeConfigCmd(dbc *Command) *cobra.Command {
+	return &cobra.Command{
+		Use:   "config",
+		Short: "print configuration",
+		Long:  "Print configuration of manager and databases",
+		Run: func(cmd *cobra.Command, args []string) {
+			dbc.PreRun(dbc, cmd, args)
+			dbc.Mgr.PrintConfig()
+		},
+	}
+}
