+++ ice/db/cmd/root.go
@@ -1,0 +1,17 @@
+package cmd
+
+import (
+	"os"
+
+	"github.com/spf13/cobra"
+)
+
+var rootCmd = &cobra.Command{
+	Use:   "db",
+	Short: "database maintenance",
+	Long:  "Database drivers, migration, status, REPL",
+	Run: func(cmd *cobra.Command, args []string) {
+		cmd.Help()
+		os.Exit(1)
+	},
+}
